refactor(contracts): extract Notes and Inbox IPC channels from ipc-channels.ts
ipc-channels.ts had grown to 937 lines — past the repo's 800-line
pre-commit ceiling. Any edit to the file (M2 needs three new calendar
channels) was blocked until it was split.

Moves the two largest blocks into sibling files:
- packages/contracts/src/notes-channels.ts (NotesChannels + two types)
- packages/contracts/src/inbox-channels.ts (InboxChannels + two types)

ipc-channels.ts re-exports both so consumers importing from
'@memry/contracts/ipc-channels' keep working unchanged. Line count drops
to 687.

No behavior change, no IPC wire-format change, no consumer changes. The
full shared test suite (1614 tests) stays green.
