+++ packages/contracts/src/ipc-channels.ts
@@ -38,125 +38,10 @@ export const VaultChannels = {
 // ============================================================================
 // Notes Channels
 // ============================================================================
+// NotesChannels lives in ./notes-channels.ts (extracted to stay under the
+// 800-line ceiling). Re-exported here so existing imports keep working.
 
-export const NotesChannels = {
-  invoke: {
-    /** Create a new note */
-    CREATE: 'notes:create',
-    /** Get a note by ID */
-    GET: 'notes:get',
-    /** Get a note by path */
-    GET_BY_PATH: 'notes:get-by-path',
-    /** Update note content/metadata */
-    UPDATE: 'notes:update',
-    /** Rename a note (changes filename) */
-    RENAME: 'notes:rename',
-    /** Move note to different folder */
-    MOVE: 'notes:move',
-    /** Delete a note */
-    DELETE: 'notes:delete',
-    /** List notes with filtering */
-    LIST: 'notes:list',
-    /** Get all tags used in notes */
-    GET_TAGS: 'notes:get-tags',
-    /** Get note links (outgoing and incoming) */
-    GET_LINKS: 'notes:get-links',
-    /** Get folder structure */
-    GET_FOLDERS: 'notes:get-folders',
-    /** Create a new folder */
-    CREATE_FOLDER: 'notes:create-folder',
-    /** Rename a folder */
-    RENAME_FOLDER: 'notes:rename-folder',
-    /** Delete a folder (recursive) */
-    DELETE_FOLDER: 'notes:delete-folder',
-    /** Check if note exists */
-    EXISTS: 'notes:exists',
-    /** Open note in external editor */
-    OPEN_EXTERNAL: 'notes:open-external',
-    /** Reveal note in file explorer */
-    REVEAL_IN_FINDER: 'notes:reveal-in-finder',
-    /** Get all property definitions (T017) */
-    GET_PROPERTY_DEFINITIONS: 'notes:get-property-definitions',
-    /** Create a property definition (T018) */
-    CREATE_PROPERTY_DEFINITION: 'notes:create-property-definition',
-    /** Update a property definition */
-    UPDATE_PROPERTY_DEFINITION: 'notes:update-property-definition',
-    /** Ensure a property definition exists (creates with defaults if missing) */
-    ENSURE_PROPERTY_DEFINITION: 'notes:ensure-property-definition',
-    /** Add an option to a select/multiselect property definition */
-    ADD_PROPERTY_OPTION: 'notes:add-property-option',
-    /** Add an option to a status property definition within a category */
-    ADD_STATUS_OPTION: 'notes:add-status-option',
-    /** Remove an option from a property definition */
-    REMOVE_PROPERTY_OPTION: 'notes:remove-property-option',
-    /** Rename an option in a property definition */
-    RENAME_PROPERTY_OPTION: 'notes:rename-property-option',
-    /** Update an option's color in a property definition */
-    UPDATE_OPTION_COLOR: 'notes:update-option-color',
-    /** Delete an entire property definition */
-    DELETE_PROPERTY_DEFINITION: 'notes:delete-property-definition',
-    /** Upload an attachment to a note (T070) */
-    UPLOAD_ATTACHMENT: 'notes:upload-attachment',
-    /** List attachments for a note */
-    LIST_ATTACHMENTS: 'notes:list-attachments',
-    /** Delete an attachment */
-    DELETE_ATTACHMENT: 'notes:delete-attachment',
-    /** Get folder config (template settings) */
-    GET_FOLDER_CONFIG: 'notes:get-folder-config',
-    /** Set folder config (template settings) */
-    SET_FOLDER_CONFIG: 'notes:set-folder-config',
-    /** Get resolved folder template (with inheritance) */
-    GET_FOLDER_TEMPLATE: 'notes:get-folder-template',
-    /** Export note as PDF (T106) */
-    EXPORT_PDF: 'notes:export-pdf',
-    /** Export note as HTML (T108) */
-    EXPORT_HTML: 'notes:export-html',
-    /** Get version history for a note (T114) */
-    GET_VERSIONS: 'notes:get-versions',
-    /** Get a specific version/snapshot (T114) */
-    GET_VERSION: 'notes:get-version',
-    /** Restore a note from a version (T114) */
-    RESTORE_VERSION: 'notes:restore-version',
-    /** Delete a specific version (T114) */
-    DELETE_VERSION: 'notes:delete-version',
-    /** Get note positions in a folder */
-    GET_POSITIONS: 'notes:get-positions',
-    /** Get all note positions */
-    GET_ALL_POSITIONS: 'notes:get-all-positions',
-    /** Reorder notes in a folder */
-    REORDER: 'notes:reorder',
-    /** Get file metadata by ID (for non-markdown files) */
-    GET_FILE: 'notes:get-file',
-    /** Resolve a WikiLink target by title (returns note or file metadata) */
-    RESOLVE_BY_TITLE: 'notes:resolve-by-title',
-    /** Get preview data for a WikiLink hover card */
-    PREVIEW_BY_TITLE: 'notes:preview-by-title',
-    /** Import files from external paths into the vault */
-    IMPORT_FILES: 'notes:import-files',
-    /** Open a file dialog to select files for import */
-    SHOW_IMPORT_DIALOG: 'notes:show-import-dialog',
-    /** Toggle local-only flag (excludes note from sync) */
-    SET_LOCAL_ONLY: 'notes:set-local-only',
-    /** Get count of local-only notes */
-    GET_LOCAL_ONLY_COUNT: 'notes:get-local-only-count'
-  },
-  events: {
-    /** Note was created (externally or internally) */
-    CREATED: 'notes:created',
-    /** Note was updated */
-    UPDATED: 'notes:updated',
-    /** Note was deleted */
-    DELETED: 'notes:deleted',
-    /** Note was renamed */
-    RENAMED: 'notes:renamed',
-    /** Note was moved */
-    MOVED: 'notes:moved',
-    /** External change detected */
-    EXTERNAL_CHANGE: 'notes:external-change',
-    /** Folder config (icon, etc.) was updated via sync */
-    FOLDER_CONFIG_UPDATED: 'notes:folder-config-updated'
-  }
-} as const
+export { NotesChannels } from './notes-channels.ts'
 
 // ============================================================================
 // Tags Channels (Tag Management & Drill-Down)
@@ -342,8 +227,7 @@ export type PropertiesInvokeChannel =
 export type VaultInvokeChannel = (typeof VaultChannels.invoke)[keyof typeof VaultChannels.invoke]
 export type VaultEventChannel = (typeof VaultChannels.events)[keyof typeof VaultChannels.events]
 
-export type NotesInvokeChannel = (typeof NotesChannels.invoke)[keyof typeof NotesChannels.invoke]
-export type NotesEventChannel = (typeof NotesChannels.events)[keyof typeof NotesChannels.events]
+export type { NotesInvokeChannel, NotesEventChannel } from './notes-channels.ts'
 
 export type TasksInvokeChannel = (typeof TasksChannels.invoke)[keyof typeof TasksChannels.invoke]
 export type TasksEventChannel = (typeof TasksChannels.events)[keyof typeof TasksChannels.events]
@@ -624,144 +508,11 @@ export type BookmarksEventChannel =
 // ============================================================================
 // Inbox Channels
 // ============================================================================
+// InboxChannels lives in ./inbox-channels.ts (extracted to stay under the
+// 800-line ceiling). Re-exported here so existing imports keep working.
 
-export const InboxChannels = {
-  invoke: {
-    // Capture operations
-    /** Capture text content */
-    CAPTURE_TEXT: 'inbox:capture-text',
-    /** Capture a URL with metadata extraction */
-    CAPTURE_LINK: 'inbox:capture-link',
-    /** Preview link metadata without creating inbox item */
-    PREVIEW_LINK: 'inbox:preview-link',
-    /** Capture an image (from drag-drop or clipboard) */
-    CAPTURE_IMAGE: 'inbox:capture-image',
-    /** Capture a voice recording */
-    CAPTURE_VOICE: 'inbox:capture-voice',
-    /** Capture a web clip (selected text from page) */
-    CAPTURE_CLIP: 'inbox:capture-clip',
-    /** Capture a PDF file */
-    CAPTURE_PDF: 'inbox:capture-pdf',
-
-    // CRUD operations
-    /** Get a single inbox item by ID */
-    GET: 'inbox:get',
-    /** List inbox items with filtering */
-    LIST: 'inbox:list',
-    /** Update an inbox item */
-    UPDATE: 'inbox:update',
-    /** Archive an inbox item (soft delete) */
-    ARCHIVE: 'inbox:archive',
-
-    // Filing operations
-    /** File an item to a folder or note */
-    FILE: 'inbox:file',
-    /** Get filing suggestions for an item */
-    GET_SUGGESTIONS: 'inbox:get-suggestions',
-    /** Track suggestion feedback (accepted/rejected) */
-    TRACK_SUGGESTION: 'inbox:track-suggestion',
-    /** Convert an item to a full note */
-    CONVERT_TO_NOTE: 'inbox:convert-to-note',
-    /** Convert an inbox item to a task */
-    CONVERT_TO_TASK: 'inbox:convert-to-task',
-    /** Link an item to an existing note */
-    LINK_TO_NOTE: 'inbox:link-to-note',
-
-    // Tag operations
-    /** Add tag to item */
-    ADD_TAG: 'inbox:add-tag',
-    /** Remove tag from item */
-    REMOVE_TAG: 'inbox:remove-tag',
-    /** Get all tags used in inbox */
-    GET_TAGS: 'inbox:get-tags',
-
-    // Snooze operations
-    /** Snooze an item */
-    SNOOZE: 'inbox:snooze',
-    /** Unsnooze an item */
-    UNSNOOZE: 'inbox:unsnooze',
-    /** Get all snoozed items */
-    GET_SNOOZED: 'inbox:get-snoozed',
-    /** Bulk snooze multiple items */
-    BULK_SNOOZE: 'inbox:bulk-snooze',
-
-    // Viewed operations (for reminder items)
-    /** Mark an inbox item as viewed */
-    MARK_VIEWED: 'inbox:mark-viewed',
-
-    // Bulk operations
-    /** Bulk file multiple items */
-    BULK_FILE: 'inbox:bulk-file',
-    /** Bulk archive multiple items */
-    BULK_ARCHIVE: 'inbox:bulk-archive',
-    /** Bulk tag multiple items */
-    BULK_TAG: 'inbox:bulk-tag',
-    /** File all stale items to unsorted */
-    FILE_ALL_STALE: 'inbox:file-all-stale',
-    // Transcription
-    /** Retry transcription for a voice item */
-    RETRY_TRANSCRIPTION: 'inbox:retry-transcription',
-
-    // Metadata
-    /** Retry metadata fetch for a link item */
-    RETRY_METADATA: 'inbox:retry-metadata',
-
-    // Stats
-    /** Get inbox statistics */
-    GET_STATS: 'inbox:get-stats',
-    /** List durable inbox jobs */
-    GET_JOBS: 'inbox:get-jobs',
-    /** Get capture patterns/insights */
-    GET_PATTERNS: 'inbox:get-patterns',
-
-    // Settings
-    /** Get stale threshold setting */
-    GET_STALE_THRESHOLD: 'inbox:get-stale-threshold',
-    /** Set stale threshold setting */
-    SET_STALE_THRESHOLD: 'inbox:set-stale-threshold',
-
-    // Archived items
-    /** List archived inbox items */
-    LIST_ARCHIVED: 'inbox:list-archived',
-    /** Unarchive an inbox item (restore to active) */
-    UNARCHIVE: 'inbox:unarchive',
-    /** Permanently delete an inbox item */
-    DELETE_PERMANENT: 'inbox:delete-permanent',
-
-    // Filing history
-    /** Get recent filing history entries */
-    GET_FILING_HISTORY: 'inbox:get-filing-history',
-
-    // Undo operations
-    /** Undo a file action (revert filedAt/filedTo/filedAction to null) */
-    UNDO_FILE: 'inbox:undo-file',
-    /** Undo an archive action (revert archivedAt to null) */
-    UNDO_ARCHIVE: 'inbox:undo-archive'
-  },
-  events: {
-    /** Item was captured */
-    CAPTURED: 'inbox:captured',
-    /** Item was updated */
-    UPDATED: 'inbox:updated',
-    /** Item was archived */
-    ARCHIVED: 'inbox:archived',
-    /** Item was filed */
-    FILED: 'inbox:filed',
-    /** Item was snoozed */
-    SNOOZED: 'inbox:snoozed',
-    /** Snoozed item became due */
-    SNOOZE_DUE: 'inbox:snooze-due',
-    /** Transcription completed */
-    TRANSCRIPTION_COMPLETE: 'inbox:transcription-complete',
-    /** Metadata fetch completed */
-    METADATA_COMPLETE: 'inbox:metadata-complete',
-    /** Processing error occurred */
-    PROCESSING_ERROR: 'inbox:processing-error'
-  }
-} as const
-
-export type InboxInvokeChannel = (typeof InboxChannels.invoke)[keyof typeof InboxChannels.invoke]
-export type InboxEventChannel = (typeof InboxChannels.events)[keyof typeof InboxChannels.events]
+export { InboxChannels } from './inbox-channels.ts'
+export type { InboxInvokeChannel, InboxEventChannel } from './inbox-channels.ts'
 
 // ============================================================================
 // Calendar Channels
